fix: use reflection for explicit interface type nil cast
In go checking for `if foo == nil`, when `foo` is an interface type is
unintuitive. If the variable is explicitly assigned, even assigning a
nil'ed pointer value of a type implementing the interface will result in
`bool(a == nil) == false`. Instead, reflection has to be used to peek
into the underlying value of the variable to check for nil.