test: added a test for the blob staleness check feature of verifier
Made the Verifier take an interface for the certVerifier instead of the struct, to allow me to test functionality that didn't depend on having a blockchain connection (like the certVerifier does).

This way we can test the blob staleness check by using a noopCertVerifier to skip all the other verification that require an eth rpc.