chore: get rid of commitmentMeta
We had commitmentMeta which consisted of commitmentMode + certVersion. This didn't really make sense to me as the certVersion should go with the cert.

So created a new struct VersionedCert which contains the certVersion and serializedCert. And then functions take this and a commitmentMode, as opposed to the serializedCert and commitmentMeta.

I believe this is a simplifying change. Its also forward looking because we will need to allow for new cert types. I plan to follow up with a cert/rbn parsing library, which will further simplify some of the logic in proxy.