chore: isolate op keccak PUT route
keccak has a very different semantic from standard and op generic commitment modes... because it writes to s3 and doesn't need to deal with eigenda certs.

It didn't make sense for it to be behind the same store interface.

There is still some cleaning up to do (left some TODO comments in the code), but this is at least a start.