refactor: server manages its own config/flags
appconfig also refactored to have an explicit high-level server config, not hidden inside EigenDA config

To fix the circular issue that happened from moving flags/config to server package, got rid of the BuildAndStartProxyServer constructor and included that as part of the main entrypoint/testbuild functions. I personally think the server should only be a server that gets injected a store, it shouldn't know how to build a store.