refactor: storeManagerBuilder
Storage configs and naming were convoluted and hard to follow. Wanted to make the builder take the whole ProxyConfig as input instead of all its individual fields. This small task ended up bubbling to this huge changeset...

Had to move the "proxyConfig" to the store in order to prevent import cycles. Made me realize the store "config" was a weird amalgation of store + secondary storage config. So broke that up, and created a separate builder package.

Also got rid of the StoreManagerBuilder class in favor of just functions. Find this cleaner personally, though I admit it wasn't strictly necessary.