feat: introduce error middleware
We were previously returning 200s on requests that should have returned 4xx/5xx errors.

Adding an error middleware to consistently deal with errors. This is a wip as there are still unstructured errors that are not being probably filtered by the error middleware right now. But we can use this framework to gradually add them and continue cleaning up our error handling (which was in a pretty bad shape).

Also refactored our middlewares to chain logging -> metrics -> errors. I might have created regressions on the metrics and logging side.. as I also cleaned up the way request mode and cert versions were propagated (previously via errors... which is just clearly wrong). These are now being propagated via a request context... open to changing if there is a better way.